feat(ssg): remove express and implement prerender for static site generation (SSG)
