+++ .gitignore
@@ -41,3 +41,6 @@ wrangler.toml
 
 # Dev
 .cursor/*
+
+# Static
+static
