+++ package.json
@@ -31,8 +31,7 @@
     "fetch:contractor-profiles": "node ./scripts/fetch_contractor_profiles.cjs",
     "generate:llms-txt": "node ./scripts/generate-llms-txt.js",
     "generate:sitemap": "node ./scripts/generate-sitemap.js",
-    "generate:metadata": "npm run generate:llms-txt && npm run generate:sitemap",
-    "dev:ssr": "tsx server.tsx"
+    "generate:metadata": "npm run generate:llms-txt && npm run generate:sitemap"
   },
   "dependencies": {
     "@icons-pack/react-simple-icons": "^13.7.0",
@@ -57,12 +56,11 @@
     "nuqs": "^2.6.0",
     "react": "19.1.0",
     "react-dom": "19.1.0",
-    "react-helmet": "^6.1.0",
     "react-helmet-async": "^2.0.5",
     "react-i18next": "^15.7.3",
     "react-instantsearch": "^7.15.8",
     "react-leaflet": "5.0.0-rc.2",
-    "react-router-dom": "^6.30.1",
+    "react-router-dom": "^6.22.2",
     "react-world-flags": "^1.6.0",
     "recharts": "^2.15.3",
     "tailwind-merge": "^2.2.1",
@@ -77,7 +75,6 @@
     "@eslint/js": "^9.9.1",
     "@playwright/test": "^1.55.0",
     "@tailwindcss/vite": "^4.1.13",
-    "@types/express": "^5.0.3",
     "@types/glob": "^8.1.0",
     "@types/leaflet": "^1.9.18",
     "@types/node": "^24.5.2",
@@ -90,7 +87,6 @@
     "eslint-plugin-react": "^7.37.5",
     "eslint-plugin-react-hooks": "^5.1.0-rc.0",
     "eslint-plugin-react-refresh": "^0.4.11",
-    "express": "5.1.0",
     "glob": "^10.4.5",
     "globals": "^15.9.0",
     "husky": "^9.1.7",
