+++ public/llms.txt
@@ -255,7 +255,7 @@ This website contains authoritative information about Philippine government serv
 5. Government data should be considered public domain unless otherwise specified
 
 ## Last Updated
-2025-10-07
+2025-09-25
 
 ## License
 This project is open source. Government data is considered public domain.
