--- src/Layout.tsx
@@ -1,10 +1,0 @@
-import { Outlet } from 'react-router-dom';
-const Layout = () => {
-  return (
-    <div>
-      <Outlet />
-    </div>
-  );
-};
-
-export default Layout;
